+++ Main_gui.py
@@ -1,3 +1,21 @@
+# SteaMidra - Steam game setup and manifest tool (SFF)
+# Copyright (c) 2025-2026 Midrag (https://github.com/Midrags)
+#
+# This file is part of SteaMidra.
+#
+# SteaMidra is free software: you can redistribute it and/or modify
+# it under the terms of the GNU General Public License as published by
+# the Free Software Foundation, either version 3 of the License, or
+# (at your option) any later version.
+#
+# SteaMidra is distributed in the hope that it will be useful,
+# but WITHOUT ANY WARRANTY; without even the implied warranty of
+# MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE.  See the
+# GNU General Public License for more details.
+#
+# You should have received a copy of the GNU General Public License
+# along with SteaMidra.  If not, see <https://www.gnu.org/licenses/>.
+
 # owner: Midrag
 import logging
 import os
@@ -63,6 +81,11 @@ def get_steam_path_gui() -> Optional[Path]:
 
 
 def main() -> None:
+    lang = get_setting(Settings.LANGUAGE)
+    if lang:
+        from sff.i18n import set_language
+        set_language(str(lang))
+
     app = QApplication(sys.argv)
     app.setApplicationName("SteaMidra")
     app.setApplicationDisplayName("SteaMidra")