+++ sff/strings.py
@@ -1,5 +1,23 @@
+# SteaMidra - Steam game setup and manifest tool (SFF)
+# Copyright (c) 2025-2026 Midrag (https://github.com/Midrags)
+#
+# This file is part of SteaMidra.
+#
+# SteaMidra is free software: you can redistribute it and/or modify
+# it under the terms of the GNU General Public License as published by
+# the Free Software Foundation, either version 3 of the License, or
+# (at your option) any later version.
+#
+# SteaMidra is distributed in the hope that it will be useful,
+# but WITHOUT ANY WARRANTY; without even the implied warranty of
+# MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE.  See the
+# GNU General Public License for more details.
+#
+# You should have received a copy of the GNU General Public License
+# along with SteaMidra.  If not, see <https://www.gnu.org/licenses/>.
+
 # owner: Midrag
-VERSION = "4.7"
+VERSION = "4.7.1"
 STEAM_WEB_API_KEY = "1DD0450A99F573693CD031EBB160907D"
 """NOTE: Not even mine lmao. Given away by oureveryday
 (source: https://github.com/SteamAutoCracks/Steam-auto-crack/issues/33)"""